+++ pkg/demoscenes/bunny.go
@@ -5,6 +5,7 @@ import . "github/chschmidt99/pt/pkg/pt"
 func Bunny(ar, fov float64) DemoScene {
 	geometry := ParseFromPath("../../assets/local/bunny/bunny.obj")
 	whiteMat := Diffuse{Albedo: NewColor(.73, .73, .73)}
+	//whiteMat := Refractive{Albedo: NewColor(.73, .73, .73), Ratio: 1.5}
 	bunny := NewSceneNode(NewMesh(geometry, whiteMat))
 	scene := NewScene()
 	scene.Add(bunny)